requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow>=9.0.0
numpy>=1.24.0
gunicorn==21.2.0